Move InputField to flat structure
