+++ packages/client/src/views/Settings/ChangePhoneModal/VerifyCodeView.tsx
@@ -18,7 +18,7 @@ import { Mutation } from 'react-apollo'
 import { changePhoneMutation } from '@client/views/Settings/mutations'
 import { get } from 'lodash'
 import { convertToMSISDN } from '@client/forms/utils'
-import { InputField } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/InputField'
 import { TextInput } from '@opencrvs/components/lib/TextInput'
 import { useSelector, useDispatch } from 'react-redux'
 import { getUserNonce, getUserDetails } from '@client/profile/profileSelectors'
