+++ packages/client/src/views/SysAdmin/Config/Application/Tabs/BirthProperties/BirthRegistrationTarget.tsx
@@ -22,7 +22,7 @@ import {
   SmallWidthInput,
   Value
 } from '@client/views/SysAdmin/Config/Application/Components'
-import { InputField } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/InputField'
 import { IStoreState } from '@client/store'
 import {
   FloatingNotification,
