+++ packages/components/src/components/DateField/DateField.tsx
@@ -12,7 +12,7 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import { ITextInputProps, IRef, TextInput } from '../TextInput/TextInput'
-import { InputLabel } from '../forms/InputField/InputLabel'
+import { InputLabel } from '../InputField/InputLabel'
 import { Omit } from '../omit'
 
 const DateWrapper = styled.div`
