+++ packages/components/src/components/InputField/InputField.tsx
@@ -13,7 +13,7 @@ import * as React from 'react'
 import styled from 'styled-components'
 import { InputError } from './InputError'
 import { InputLabel } from './InputLabel'
-import { colors } from '../../colors'
+import { colors } from '../colors'
 const InputHeader = styled.div`
   display: flex;
   justify-content: space-between;
