+++ packages/components/src/components/InputField/index.ts
@@ -9,7 +9,6 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './PDFViewer/PDFViewer'
+export * from './InputError'
+export * from './InputLabel'
+export * from './InputField'
