+++ packages/components/src/components/Radio/RadioGroup.tsx
@@ -12,7 +12,7 @@
 import * as React from 'react'
 import { RadioButton } from '../interface/RadioButton'
 import { NoticeWrapper } from '../DateField'
-import { InputLabel } from '../forms/InputField/InputLabel'
+import { InputLabel } from '../InputField/InputLabel'
 import styled from 'styled-components'
 
 const Wrapper = styled.div`
