+++ packages/components/src/components/interface/DataTable/SortAndFilter.tsx
@@ -10,7 +10,6 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import * as React from 'react'
-import { InputField, IInputFieldProps } from '../../forms'
 import { Omit } from '../../omit'
 import styled from 'styled-components'
 import {
