+++ packages/components/src/components/interface/LocationSearch/LocationSearch.tsx
@@ -13,7 +13,7 @@ import * as React from 'react'
 import styled from 'styled-components'
 import { Location } from '../../icons'
 import { PrimaryButton } from '../../buttons'
-import { InputError } from '../../forms/InputField/InputError'
+import { InputError } from '../../InputField/InputError'
 
 const SEARCH_DEBOUNCE_DURATION = 300
