+++ packages/login/src/views/StepOne/StepOneForm.tsx
@@ -14,7 +14,7 @@ import { WrappedComponentProps as IntlShapeProps, injectIntl } from 'react-intl'
 import styled from 'styled-components'
 import { Form, Field } from 'react-final-form'
 import { PrimaryButton, LinkButton } from '@opencrvs/components/lib/buttons'
-import { InputField, THEME_MODE } from '@opencrvs/components/lib/forms'
+import { InputField, THEME_MODE } from '@opencrvs/components/lib/InputField'
 import { TextInput } from '@opencrvs/components/lib/TextInput'
 import { ErrorMessage } from '@opencrvs/components/lib/ErrorMessage'
 import { PasswordInput } from '@opencrvs/components/lib/PasswordInput'
