+++ packages/login/src/views/resetCredentialsForm/phoneNumberVerificationForm.tsx
@@ -18,7 +18,7 @@ import {
 import { authApi } from '@login/utils/authApi'
 import { phoneNumberFormat } from '@login/utils/validate'
 import { PrimaryButton } from '@opencrvs/components/lib/buttons'
-import { InputField } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/InputField'
 import { TextInput } from '@opencrvs/components/lib/TextInput'
 import { ActionPageLight } from '@opencrvs/components/lib/interface'
 import * as React from 'react'
